pyproject.toml
[project]
name = "inh"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "scikit-image>=0.25.2",
    "streamlit>=1.50.0",
    "streamlit-drawable-canvas>=0.9.3",
]
